lib: internal: cacheimpl: PrePackageRecord: set offset type to uint32_t
Fixes compiling on 64-bit platforms.

+++ cpp/lib/src/internal/cacheimpl.hpp
@@ -46,7 +46,7 @@ class CacheImpl
  public:
 	struct PrePackageRecord
 	{
-		size_t offset;
+		uint32_t offset;
 		const pair< shared_ptr< const ReleaseInfo >, shared_ptr< File > >* releaseInfoAndFile;
 	};
  private:
